move dsm tests to test_tracing.py
